Eliminate the dependencies on sourceSets.test.output for metadata
