+++ build.gradle
@@ -1141,7 +1141,7 @@ project(':core') {
     testImplementation testFixtures(project(':clients'))
     testImplementation testFixtures(project(':group-coordinator'))
     testImplementation testFixtures(project(':share-coordinator'))
-    testImplementation project(':metadata').sourceSets.test.output
+    testImplementation testFixtures(project(':metadata'))
     testImplementation project(':raft').sourceSets.test.output
     testImplementation testFixtures(project(':server-common'))
     testImplementation testFixtures(project(':storage:storage-api'))
@@ -1399,6 +1399,8 @@ project(':metadata') {
     archivesName = "kafka-metadata"
   }
 
+  apply plugin: 'java-test-fixtures'
+
   configurations {
     generator
   }
@@ -1413,6 +1415,9 @@ project(':metadata') {
     implementation libs.slf4jApi
 
     testImplementation testLog4j2Libs
+    testFixturesImplementation testFixtures(project(':server-common'))
+    testFixturesImplementation libs.junitJupiter
+
     testImplementation libs.junitJupiter
     testImplementation libs.jqwik
     testImplementation libs.mockitoCore
@@ -3522,8 +3527,8 @@ project(':jmh-benchmarks') {
     implementation project(':connect:json')
     implementation testFixtures(project(':clients'))
     implementation testFixtures(project(':server-common'))
+    implementation testFixtures(project(':metadata'))
     implementation project(':server').sourceSets.test.output
-    implementation project(':metadata').sourceSets.test.output
 
     implementation libs.jmhCore
     annotationProcessor libs.jmhGeneratorAnnProcess
@@ -4091,6 +4096,7 @@ gradle.projectsEvaluated {
     'storage/api': ':storage:storage-api',
     'group-coordinator': ':group-coordinator',
     'share-coordinator': ':share-coordinator',
+    'metadata': ':metadata',
   ]
   allprojects { proj ->
     proj.configurations.all { config ->
