+++ core/src/test/scala/unit/kafka/security/authorizer/AuthorizerTest.scala
@@ -34,8 +34,7 @@ import org.apache.kafka.common.resource.{PatternType, ResourcePattern, ResourceP
 import org.apache.kafka.common.security.auth.{KafkaPrincipal, SecurityProtocol}
 import org.apache.kafka.common.utils.internals.{SecurityUtils => JSecurityUtils}
 import org.apache.kafka.controller.MockAclMutator
-import org.apache.kafka.metadata.authorizer.StandardAuthorizer
-import org.apache.kafka.metadata.authorizer.StandardAuthorizerTest.AuthorizerTestServerInfo
+import org.apache.kafka.metadata.authorizer.{AuthorizerTestServerInfo, StandardAuthorizer}
 import org.apache.kafka.security.authorizer.AclEntry.{WILDCARD_HOST, WILDCARD_PRINCIPAL_STRING}
 import org.apache.kafka.server.authorizer._
 import org.apache.kafka.server.util.ServerTestUtils
