+++ metadata/src/testFixtures/java/org/apache/kafka/metadata/RecordTestUtils.java
